Pass grace-period to virt-controller when stopping or restarting VM
The grace-period was never getting passed to virt-controller which
deletes the VMI, causing Delete() to use the default TerminationGracePeriodSeconds
in the VMI Spec. This is particularly a problem with Windows VMIs, which
by default have a TerminationGracePeriodSeconds of 3600 seconds (one hour).
If the user wants to force stop/restart a VM with grace-period 0, we should honor
this and delete the VMI with that grace-period.

Signed-off-by: prnaraya <prnaraya@redhat.com>