Have libvirt Domain manager KillVMI when signalling shutdown for grace-period=0
When kubectl delete vm --force --grace-period=0 is called, it requests deletion of the VM
and the VMI, placing the grace-period=0 in the ObjectMeta DeletionGracePeriodSeconds.
Before signalling graceful shutdown, we first check if DeletionGracePeriodSeconds is 0.
If it is, then a force deletion was requested and we want to honor the grace-period of 0,
so we will kill the VMI instead.

Signed-off-by: prnaraya <prnaraya@redhat.com>