VMI GracePeriodSeconds converter now uses VMI.DeletionGracePeriodSeconds
Previously, the converter which is used by the libvirt domain manager used
the value in VMI.Spec.TerminationGracePeriodSeconds. This was under the
assumption that this value would be updated when the deletion was  invoked.
However, testing demonstrated that this value does not end up getting updated,
resulting in VMIs with a high TerminationGracePeriodSeconds will end up
getting stuck when attempting to force delete with a grace period of 0,
as the TerminatingGracePeriodSeconds value is not getting updated. We modify
the converter to instead use the VMI's DeletionGracePeriodSeconds, which is
set once the deletion is signalled. The remainining check to the TerminatingGracePeriodSeconds
in the VMI Spec is in the (unlikely) case there is no DeletionGracePeriodSeconds
set, the original TerminatingGracePeriodSeconds is not lost.

Signed-off-by: prnaraya <prnaraya@redhat.com>