virt-launcher: update grace period in metadata cache when signaling VMI shutdown
Grace period converter should initially check the VMI's ObjectMeta for DeletionGracePeriodSeconds,
otherwise use the TerminationGracePeriodSeconds in the VMI Spec if it exists, or use the default
grace period. When signaling VMI shutdown, in addition to setting the DeletionTimestamp in the metadata
cache if it hasn't already, also set the DeletionGracePeriodSeconds.

Signed-off-by: prnaraya <prnaraya@redhat.com>